+++ R/argument_convention.R
@@ -1,0 +1,12 @@
+#' Standard Arguments
+#'
+#' The documentation to this function lists all the arguments in `staged.dependencies`
+#' that are used repeatedly.
+#'
+#' @param verbose (`numeric`) \cr verbosity level, incremental;
+#'   (0: None, 1: packages that get installed + high-level git operations,
+#'   2: includes git checkout infos)
+#'
+#' @keywords internal
+#' @name argument_convention
+NULL